refactor: simplified AddRegionModal
