+++ src/pages/MapPage/components/RegionDrawing/AddRegionModal.tsx
@@ -1,14 +1,16 @@
-import React, { useState, useRef } from "react";
+import React, { useState, useRef, useMemo } from "react";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { MapContainer, TileLayer, FeatureGroup } from "react-leaflet";
 import { EditControl } from "react-leaflet-draw";
 import { useDispatch } from "react-redux";
-import { Square, Circle, Pentagon, X, Backpack } from "lucide-react";
 import { addRegion } from "@/redux/regionActions";
 import { Region } from "@/types/region";
 import "leaflet/dist/leaflet.css";
 import "leaflet-draw/dist/leaflet.draw.css";
+import { drawOptions } from "./drawOptions";
+import { generateRandomColor } from "@/utils/colors";
+import { calculatePolygonArea } from "@/utils/map";
 
 interface AddRegionModalProps {
     onClose: () => void;
@@ -21,39 +23,6 @@ function AddRegionModal({ onClose }: AddRegionModalProps) {
         [number, number][]
     >([]);
     const featureGroupRef = useRef<any>(null);
-
-    // Generate random color for new regions
-    const generateRandomColor = () => {
-        const colors = [
-            "#FF6B6B",
-            "#4ECDC4",
-            "#45B7D1",
-            "#96CEB4",
-            "#FFEAA7",
-            "#DDA0DD",
-            "#98D8C8",
-        ];
-        return colors[Math.floor(Math.random() * colors.length)];
-    };
-
-    // Calculate area of polygon (simplified calculation)
-    const calculatePolygonArea = (coordinates: [number, number][]) => {
-        if (coordinates.length < 3) return 0;
-
-        let area = 0;
-        const n = coordinates.length;
-
-        for (let i = 0; i < n; i++) {
-            const j = (i + 1) % n;
-            area += coordinates[i][0] * coordinates[j][1];
-            area -= coordinates[j][0] * coordinates[i][1];
-        }
-
-        area = Math.abs(area) / 2;
-        area = area * 12391; // Rough conversion to km²
-
-        return Math.round(area * 100) / 100;
-    };
 
     // Handle shape creation
     const handleCreated = (e: any) => {
@@ -135,43 +104,15 @@ function AddRegionModal({ onClose }: AddRegionModalProps) {
     };
 
     // Get draw options based on selected mode
-    const getDrawOptions = () => ({
-        polyline: false,
-        polygon: {
-            allowIntersection: false,
-            drawError: {
-                color: "#e1e100",
-                message: "<strong>Oh snap!<strong> you can't draw that!",
-            },
-            shapeOptions: {
-                color: "#97009c",
-            },
-        },
-        rectangle: {
-            shapeOptions: {
-                //   clickable: false,
-            },
-        },
-        circle: {
-            shapeOptions: {
-                color: "#662d91",
-            },
-        },
-        marker: false,
-        circlemarker: false,
-    });
 
     return (
         <div className="absolute inset-0 flex items-center justify-center z-50">
             <div className="relative bg-white rounded-lg w-full h-full overflow-hidden flex">
-                {/* Left Panel - Controls */}
+                {/* Controls */}
                 <div className="absolute right-0 bottom-0 w-80 p-6 border-r border-gray-200 rounded-lg bg-white flex flex-col z-100">
                     <div>
                         {currentCoordinates.length > 0 && (
                             <div className="text-sm text-green-600 bg-green-50 p-3 rounded">
-                                ✓ Shape drawn! {currentCoordinates.length}{" "}
-                                points
-                                <br />
                                 Area: ~
                                 {calculatePolygonArea(currentCoordinates)} km²
                             </div>
@@ -213,7 +154,7 @@ function AddRegionModal({ onClose }: AddRegionModalProps) {
                     </div>
                 </div>
 
-                {/* Right Panel - Map */}
+                {/* Map */}
                 <div className="absolute z-50 w-full h-full">
                     <MapContainer
                         center={[52.52, 13.405]}
@@ -232,7 +173,7 @@ function AddRegionModal({ onClose }: AddRegionModalProps) {
                                 onCreated={handleCreated}
                                 onEdited={handleEdited}
                                 onDeleted={handleDeleted}
-                                draw={getDrawOptions()}
+                                draw={drawOptions}
                                 edit={{
                                     remove: true,
                                     edit: {},
